==> 02_notebooks/04_propensity_model.ipynb <==
# Load Segmented Dataset

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import roc_auc_score, classification_report

pd.set_option("display.max_columns", None)

df = pd.read_csv("../01_data/processed/credit_card_segmented.csv")

df.shape


# Define Features

features = [
    "engagement_score",
    "utilization_proxy",
    "avg_bill_amt",
    "default"
]

X = df[features]
y = df["adopted"]

# Train Data

X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.3,
    random_state=42
)


# Fit Regression

model = LogisticRegression(max_iter=1000)
model.fit(X_train, y_train)

# Evaluate Model Performance

y_pred_prob = model.predict_proba(X_test)[:, 1]

auc = roc_auc_score(y_test, y_pred_prob)
print(f"AUC: {round(auc, 2)}")

# Interpret Coefficients

coef_df = pd.DataFrame({
    "feature": features,
    "coefficient": model.coef_[0]
}).sort_values("coefficient", ascending=False)

coef_df


# Generate Propensity Scores for Customers

df["propensity_score"] = model.predict_proba(X)[:, 1]

df["propensity_score"].describe()


# Rank Customers

df = df.sort_values("propensity_score", ascending=False)
df[["customer_segment", "propensity_score", "adopted"]].head(10)


## Model Summary

The logistic regression model is built to determine the likelihood probability of a customer adopting the 5% cash-back offer.

The logistic regression model achieved an AUC of approximately 0.73, indicating a relatively strong ranking capability as the model can correctly rank 
an adopter higher than a non-adopter about 73% of the time.

- Key Factors
    - Higher engagement -> Increased adoption likelihood
    - Higher utilization -> Decrease adoption likelihood
    - Default history -> Decrease adoption likelihood

# Save Dataset

df.to_csv("../01_data/processed/credit_card_with_propensity.csv", index=False)
